refactor: Phase 8 — extract 3 composables from ChatView, fix TDZ crash
- useSessionManagement: session create/switch/delete, name editing, persistence (-203 lines)
- useMessageActions: message delete/hide, edit save/cancel, branch, image regen (-194 lines)
- useChatGeneration: sendMessage, startGeneration, image-gen lifecycle (-152 lines)
- Fix TDZ crash: move const t before composable calls
- Clean up 15+ unused imports after extraction
- ChatView: 3767 -> 2994 lines (-20.5%)
- Update REFACTOR_PLAN, Roadmap, ARCHITECTURE
